Initial commit of hill-type model
This is prior to the attempt to incorporate shortening induced force
depression and so is a safe reversion point to a 'default' hill type
model. However, this is after the re-derivation of the CE velocity
calculation to incorporate a f_st intended to be used to incorporate
force depression. It is just set to zero currently.